fixes dependencies
